Fixed decoder reset behaviour on input timeout.
Fixed decoder behaviour that would cause the complete loss of all data
followin a single input timeout, even after another complete and
correct preamble. The problem was solved by not resetting the timer
each time a timeout would occur.

Conected the decoding function to the reception complete event.

The decoder was tested for the first time with a 12V DCC signal.
GPIO A and B are working successfully.